Creating array to store secrets
I'm creating an array to store the secrets, but I'm getting an error saying 'Cannot read property length of undefined'.  I reckon this means that I cannot do this because there is nothing in the array yet.  I have to figure out how to create an array and store a secret after a user has signed up