Added 1st joke
Had a lot of difficulty trying to figure out why the secrets page was showing the undefined length error.  I referred to John H notes, and got the idea to try date/timestamp.  This didn't solve it.  I added a for loop in my jade page, this didn't solve it either.  I then went through my code and JH's code line by line and really couldn't see a difference, until I went to line 87 to 92 in my code.  I then realised that I was not sending in the array in this res.render function. Problem now solved