Secrets are now storing in local storage
Originally, when I stopped the server and restarted it again, the secrets were there in the secrets page, but as I added a new one, the previous ones were deleted.  This was because I was actually creating a new array each time I stopped the server and restarted dit and added a new secret.  So, what I had to do was, I had to pull the original array from local storage, and add the new secret to that.  I think I've gone down a bit of a rabbit hole, but it seems to work.